python_simulation/requirements.txt
absl-py==2.1.0
chex==0.1.86
contourpy==1.2.0
cycler==0.12.1
equinox==0.11.7
etils==1.9.4
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.3
ml_dtypes==0.5.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
opt_einsum==3.4.0
optax==0.2.3
packaging==23.2
pandas==2.2.1
pillow==10.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
scipy==1.12.0
six==1.16.0
sympy==1.12
toolz==0.12.1
tqdm==4.66.2
triton==2.2.0
typeguard==2.13.3
typing_extensions==4.10.0
tzdata==2024.1
